requirements.txt
# 深度学习框架
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0

# 数据处理
numpy>=1.24.0
pandas>=2.0.0
pillow>=10.0.0
scikit-learn>=1.3.0

# 可视化和日志
matplotlib>=3.7.0
seaborn>=0.12.0
tensorboard>=2.13.0

# 配置管理
pyyaml>=6.0

# 评估指标
scikit-image>=0.21.0
tqdm>=4.65.0
